% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wstats-package.R
\docType{package}
\name{wstats-package}
\alias{wstats}
\alias{wstats-package}
\title{wstats: Weighted Descriptive Statistics}
\description{
Weighted versions of common descriptive statistics (variance, standard deviation, covariance, correlation, quantiles) for importance weights. Computationally intensive routines are implemented in C++ via cpp11.
}
\author{
\strong{Maintainer}: Johann D. Gaebler \email{me@jgaeb.com}

}
\keyword{internal}
